packages/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic-settings==2.7.1
PyJWT==2.10.1
cryptography==44.0.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
pytest==8.3.4
httpx==0.28.1
